Escape Unicode surrogates in logged path specifications
