+++ plaso/cli/status_view.py
@@ -2,6 +2,7 @@
 """The status view."""
 
 import ctypes
+import re
 import sys
 import time
 
@@ -34,6 +35,8 @@ class StatusView(object):
           'storage media device'),
       dfvfs_definitions.SOURCE_TYPE_STORAGE_MEDIA_IMAGE: (
           'storage media image')}
+
+  _UNICODE_SURROGATES_RE = re.compile('[\ud800-\udfff]')
 
   _UNITS_1024 = ['B', 'KiB', 'MiB', 'GiB', 'TiB', 'EiB', 'ZiB', 'YiB']
 
@@ -232,6 +235,25 @@ class StatusView(object):
       self._PrintEventsStatus(processing_status.events_status)
 
     self._output_writer.Write('\n')
+
+  def _GetPathSpecificationString(self, path_spec):
+    """Retrieves a printable string representation of the path specification.
+
+    Args:
+      path_spec (dfvfs.PathSpec): path specification.
+
+    Returns:
+      str: printable string representation of the path specification.
+    """
+    path_spec_string = path_spec.comparable
+
+    if self._UNICODE_SURROGATES_RE.search(path_spec_string):
+      path_spec_string = path_spec_string.encode(
+          'utf-8', errors='surrogateescape')
+      path_spec_string = path_spec_string.decode(
+          'utf-8', errors='backslashreplace')
+
+    return path_spec_string
 
   def _PrintAnalysisStatusUpdateLinear(self, processing_status):
     """Prints an analysis status update in linear mode.
@@ -502,7 +524,8 @@ class StatusView(object):
             ''])
         self._output_writer.Write(output_text)
         for path_spec in processing_status.error_path_specs:
-          self._output_writer.Write(path_spec.comparable)
+          path_spec_string = self._GetPathSpecificationString(path_spec)
+          self._output_writer.Write(path_spec_string)
           self._output_writer.Write('\n')
 
     self._output_writer.Write('\n')
